Extract the base protocol so that we can reuse for meta
