+++ lib/dalli.rb
@@ -62,6 +62,7 @@ require_relative 'dalli/key_manager'
 require_relative 'dalli/pipelined_getter'
 require_relative 'dalli/ring'
 require_relative 'dalli/protocol'
+require_relative 'dalli/protocol/base'
 require_relative 'dalli/protocol/binary'
 require_relative 'dalli/protocol/connection_manager'
 require_relative 'dalli/protocol/response_buffer'
